Refactor download function
