+++ src/utils/format.js
@@ -1,4 +1,5 @@
 /* eslint-disable no-useless-escape */
+import _ from 'lodash';
 
 /**
  * Format hideout object to .hideout style
@@ -20,14 +21,38 @@ function formatHideoutObject(object = {}) {
 }
 
 /**
- * Match img tag to array from form content
+ * Match img tag to array from form content in create page
  * @param {string} content
  */
 function formatImgTagFromContent(content = '') {
   return String(content).match(/<img\s[^>]*?src\s*=\s*['\"]([^'\"]*?)['\"][^>]*?>/g) || [];
 }
 
+/**
+ * Format hideout file from database fileContent
+ * @param {string} content
+ */
+function formatHideoutFromFileContent(content = '') {
+  // Rebuild Constructor
+  try {
+    const { Objects, ...Args } = JSON.parse(content);
+    const Title = Object.keys(Args).map(o => {
+      return (`${o} = ${o === 'Hideout Hash' ? _.get(Args, o) : JSON.stringify(_.get(Args, o))}\n`);
+    })
+    const Files = Objects.map(o => {
+      const { Name, ...args } = o;
+      return `${Name} = ${formatHideoutObject(args)}\n`;
+    });
+    // Combine hideout object
+    return [...Title, ['\n'], ...Files].join('');
+  } catch (e) {
+    console.warn('formatHideoutFromFileContent', e);
+    return '';
+  }
+}
+
 export {
   formatHideoutObject,
   formatImgTagFromContent,
+  formatHideoutFromFileContent,
 }
